+++ PyIRTAM/tests/__init__.py
@@ -1,0 +1,1 @@
+"""Initalization file for unit and integration tests."""